cybersecurity-chatbot/requirements.txt
torch>=2.0.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.0.2
matplotlib>=3.5.0
seaborn>=0.12.0
pyyaml>=6.0.0
pytest>=7.0.0
black>=22.0.0
pylint>=2.15.0